[atlas-gradle-plugin] tpatch manifest only contains user new components
